fix: находки второго код-ревью — деление надбавок, потеря чека, утечка токена
Деление позиций (все три платформы)
- splitByWeight раздавал остаток от округления участнику с НУЛЕВЫМ весом:
  у proportional-надбавки вес равен базовой доле, и тот, кто ничего не ел,
  выигрывал tie-break по меньшему userId и платил весь сервисный сбор.
  Инвариант «сумма долей == итог» при этом сходился, поэтому ошибка была
  невидимой. Схлопываем дубли по id и отбрасываем нулевые веса до деления.
- Клиентские зеркала не получили += из прошлого ревью: iOS и Android
  затирали долю дубля вместо накопления, из-за чего derivedShares возвращал
  nil и сохранение блокировалось «баг расчёта» без выхода.

Потеря данных
- Правка неотправленной офлайн-записи стирала чек: ветка editEntry не
  восстанавливала draftItems, и в outbox уходил items=nil — при синке
  создавался плоский расход без позиций.
- Голосовая правка без позиций в ответе («платил Саша») стирала показанный
  чек и в том же вызове снимала undo.
- Смена группы сохраняла позиции прежней: сохранение шло itemized-веткой с
  чужими userId.
- Плоский черновик с позициями без цен обнулял распознанную сумму
  («пицца и салат, всего 1200» → 0).
- OutboxStore не отличал «файла нет» от «не прочитался»: под
  completeFileProtection фоновый запуск на залоченном устройстве давал
  пустую очередь, и первая же запись затирала неотправленные расходы.
- Двойной тап «Сохранить» офлайн создавал два расхода: isSaving
  выставлялся ПОСЛЕ офлайн-веток, а localId был новым на каждый вызов.

Безопасность и запуск
- Dockerfile из прошлого ревью перестал класть conf/ в рантайм-образ, а
  initI18n вызывается безусловно и паникует на отсутствующей директории —
  контейнер не поднимался вообще.
- tgGet отдавал ошибки *url.Error как есть: bot-токен лежит в ПУТИ
  (/bot<token>/getFile), stripPassword его не трогает, и любой таймаут
  сливал токен в логи.
- Write-path не проверял kind/split и не ограничивал число позиций/долей
  от клиента: произвольная строка ложилась в базу, неизвестный split молча
  трактовался как equally.
- sanitizeDraft не ограничивал цену/вес сверху: сумма позиций переполняла
  int, а iOS трапится на amount*weight.

Прочее
- abs(Int.MIN_VALUE) отрицателен — крупнейший долг уезжал в конец списка.
- Обрыв записи системой (входящий звонок) молча замораживал UI до автостопа
  и слал в parse обрезанный WAV; теперь экран останавливает запись.
- Отказ в доступе к микрофону в VoiceOver-режиме оставлял замок и оверлей
  поверх формы — экран становился полностью нежатым.

Проверка: go vet + go test ./internal/..., Android testDebugUnitTest +
lintDebug + verifyRoborazziDebug, iOS SplittyTests (139). Новые тесты
проверены на падение без фикса.

+++ android/app/src/main/java/com/zagir/splitty/core/money/Money.kt
@@ -66,7 +66,10 @@ fun aggregateByCurrency(amounts: List<CurrencySum>): List<CurrencySum> {
             )
         }
         .sortedWith(
-            compareByDescending<CurrencySum> { abs(it.sum) }.thenBy { it.currency }
+            // abs на Long, а не на Int: насыщение выше может дать ровно
+            // Int.MIN_VALUE, а abs(Int.MIN_VALUE) отрицателен — крупнейший долг
+            // уезжал в конец списка, и «основной» показывалась не та валюта
+            compareByDescending<CurrencySum> { abs(it.sum.toLong()) }.thenBy { it.currency }
         )
 }
